Update test/acceptance/fixture/metadataDirs/full/ [+]
Replaced with files generated in test/acceptance/tmp/metadata/ from
new intake command.  The `intake` acceptance test was failing due
to metadata files previously generated by Perl legacy scripts
formatting the JSON in a different way, with some differences in how
"\n", whitespace, and entity references are handled in the strings.
(Note that test/acceptance/expected/metadata-from-intake/ is just a
symlink to test/acceptance/fixture/metadataDirs/full/, which is
why it's necessary to replace this directory)

Used temporary normalization functions to get the new intake
command output generated by the test to match the old expected
JSON files using `diff -w -r`.  Doesn't seem necessary to get
whitespace to be equivalent because whitespace in JSON structure
is not significant, and since the whitespace differences in the
property values occurred only in description and description_html,
which are displayed in HTML in the browser and thus will have all
whitespace condensed to a single space at render-time, it doesn't
seem worth it to match on `diff -r`.