test/unit/: add test for Book.authors with snapshots -- SNAPSHOTS NOT YET COMPLETE [+]
All authors snapshots checked in with this commit have been confirmed to be correct.

Snapshots are missing for the following:

* 9780814718803
* 9780814750957

See https://jira.nyu.edu/jira/browse/NYUP-733 for details.

Also, the Supafolio record for EPUB 9780814735336 appears to have an error:

      "name": "Robert L. Hayman Jr.",
      "first": "Robert  L. Hayman",
      "last": "Jr.",

After this is corrected in Supafolio will need regenerate this snapshot.