test/acceptance/fixture/supafolio-api/: replace the symlinks to *hacked* files with the actual files and add README files explaining why this was necessary [+]
9780814750957.json.README:

This actually the Supafolio API response for 9780814752739 but with isbn13 changed to 9780814750957.  See https://jira.nyu.edu/jira/browse/NYUP-733 for the reason why this is necessary.

Originally this file was a symlink to 9780814750957.json.hacked-9780814752739, but unless --preserve-symlinks flag is set Node will apparently not load the file as JSON, which can cause parsing to fail.